Fix: "형변환 에러 수정" Id값 모두 Long형으로 통일

+++ src/main/java/com/app/domain/categorizedsummary/repository/CategorizedSummaryRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategorizedSummaryRepository extends JpaRepository<CategorizedSummary, Long> {
-    boolean existsByCategoryCategoryIdAndSummarySummaryId(Long categoryId, Integer summaryId);
+    boolean existsByCategoryCategoryIdAndSummarySummaryId(Long categoryId, Long summaryId);
 
-    boolean existsBySummarySummaryId(Integer summaryId);
+    boolean existsBySummarySummaryId(Long summaryId);
 
     Page<CategorizedSummary> findByCategoryCategoryId(Long categoryId, Pageable pageable);
 }
