+++ src/main/java/com/app/domain/summary/repository/SummaryRepository.java
@@ -3,6 +3,6 @@ package com.app.domain.summary.repository;
 import com.app.domain.summary.entity.Summary;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SummaryRepository extends JpaRepository<Summary, Integer> {
+public interface SummaryRepository extends JpaRepository<Summary, Long> {
 
 }
